refactor: add factory and default-filters constant in  side-filter template
