+++ src/components/UI/template/SideFilter/factory/index.ts
@@ -1,0 +1,1 @@
+export * from './factory.component'
